refactor(CampagneSelector): extract CampagneSelector in component
